ACC-2274 fix renaming content -> contentstore
Fixes rebase with main after #106 got merged
