+++ contentgrid-appserver-autoconfigure/build.gradle
@@ -9,10 +9,10 @@ dependencies {
     implementation platform(project(':contentgrid-appserver-platform'))
 
     compileOnly project(':contentgrid-appserver-application-model')
-    compileOnly project(':contentgrid-appserver-content-api')
-    compileOnly project(':contentgrid-appserver-content-impl-encryption')
-    compileOnly project(':contentgrid-appserver-content-impl-fs')
-    compileOnly project(':contentgrid-appserver-content-impl-s3')
+    compileOnly project(':contentgrid-appserver-contentstore-api')
+    compileOnly project(':contentgrid-appserver-contentstore-impl-encryption')
+    compileOnly project(':contentgrid-appserver-contentstore-impl-fs')
+    compileOnly project(':contentgrid-appserver-contentstore-impl-s3')
     compileOnly project(':contentgrid-appserver-domain')
     compileOnly project(':contentgrid-appserver-json-schema')
     compileOnly project(':contentgrid-appserver-query-engine-api')
@@ -28,10 +28,10 @@ dependencies {
     testImplementation 'org.springframework.boot:spring-boot-starter-web'
 
     testImplementation project(':contentgrid-appserver-application-model')
-    testImplementation project(':contentgrid-appserver-content-api')
-    testImplementation project(':contentgrid-appserver-content-impl-encryption')
-    testImplementation project(':contentgrid-appserver-content-impl-fs')
-    testImplementation project(':contentgrid-appserver-content-impl-s3')
+    testImplementation project(':contentgrid-appserver-contentstore-api')
+    testImplementation project(':contentgrid-appserver-contentstore-impl-encryption')
+    testImplementation project(':contentgrid-appserver-contentstore-impl-fs')
+    testImplementation project(':contentgrid-appserver-contentstore-impl-s3')
     testImplementation project(':contentgrid-appserver-domain')
     testImplementation project(':contentgrid-appserver-json-schema')
     testImplementation project(':contentgrid-appserver-query-engine-api')
