+++ contentgrid-appserver-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,6 +1,6 @@
-com.contentgrid.appserver.autoconfigure.content.EncryptedContentStoreAutoConfiguration
-com.contentgrid.appserver.autoconfigure.content.FileSystemContentStoreAutoConfiguration
-com.contentgrid.appserver.autoconfigure.content.S3ContentStoreAutoConfiguration
+com.contentgrid.appserver.autoconfigure.contentstore.EncryptedContentStoreAutoConfiguration
+com.contentgrid.appserver.autoconfigure.contentstore.FileSystemContentStoreAutoConfiguration
+com.contentgrid.appserver.autoconfigure.contentstore.S3ContentStoreAutoConfiguration
 com.contentgrid.appserver.autoconfigure.domain.ContentGridDomainAutoConfiguration
 com.contentgrid.appserver.autoconfigure.flyway.FlywayPostgresAutoConfiguration
 com.contentgrid.appserver.autoconfigure.json.schema.ApplicationResolverAutoConfiguration
